requirements.txt
APScheduler==3.5.3
certifi==2018.11.29
chardet==3.0.4
defusedxml==0.5.0
diff-match-patch==20181111
Django==2.1.7
django-advanced-reports==0.9.24
django-archive==0.1.5
django-common-helpers==0.9.2
django-cron==0.5.1
django-crontab==0.7.1
django-import-export==1.2.0
django-report-builder==6.2.2
django-widget-tweaks==1.4.3
djangorestframework==3.9.2
et-xmlfile==1.0.1
idna==2.8
jdcal==1.4
odfpy==1.4.0
openpyxl==2.6.1
psycopg2==2.7.7
Pygments==2.3.1
python-dateutil==2.8.0
pytz==2018.9
PyYAML==3.13
requests==2.21.0
six==1.12.0
tablib==0.12.1
tzlocal==1.5.1
unicodecsv==0.14.1
urllib3==1.24.1
xlrd==1.2.0
xlwt==1.3.0
